following the test structure properly now
